+++ client/components/discountDetail/DiscountDetailForm.jsx
@@ -4,6 +4,7 @@ import CSSModules from 'react-css-modules';
 import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
 import TextField from '../common/TextField.jsx';
+import DiscountValueField from './DiscountValueField.jsx';
 
 import styles from './discountDetail.css';
 
@@ -12,6 +13,11 @@ const DiscountDetailForm = ({ discount, items, errors, onChange }) => {
   const discountOptions = [
     { name: 'applyTypeID', value: 0, label: 'Everything' },
     { name: 'applyTypeID', value: 1, label: 'Specific Items' }
+  ];
+
+  const valueType = [
+    { name: 'discountTypeID', value: 0, label: 'Percentage' },
+    { name: 'discountTypeID', value: 1, label: 'Currency Value' }
   ];
 
   const displayItems = items.map(item => {
@@ -23,6 +29,7 @@ const DiscountDetailForm = ({ discount, items, errors, onChange }) => {
   });
 
   const selectedDiscountOption = discountOptions.find(option => option.value === discount.applyTypeID);
+  const selectedValueType = valueType.find(option => option.value === discount.discountTypeID);
   const displayItemSelection = selectedDiscountOption && !!selectedDiscountOption.value;
 
   return (
@@ -38,19 +45,37 @@ const DiscountDetailForm = ({ discount, items, errors, onChange }) => {
               onChange={onChange}
               errorText={errors.discountName} />
           </div>
+          <div className={styles['discount-option-control']}>
+            <div className="row">
+              <div className={styles['discount-name-control']}>
+                <SelectField
+                  floatingLabelText="Value Type"
+                  value={selectedValueType}
+                  onChange={onChange}
+                >
+                  {valueType.map(option => {
+                    return (
+                      <MenuItem
+                        key={option.value}
+                        primaryText={option.label}
+                        value={option} />
+                    );
+                  })}
+                </SelectField>
+              </div>
+            </div>
+          </div>
           <div className="row">
             <div>
-              <TextField
+              <DiscountValueField
                 fullWidth
+                isPercent={selectedValueType.value === 0}
                 name="value"
                 value={discount.value}
-                floatingLabelText="Discount Rate"
+                floatingLabelText="Value"
                 onChange={onChange}
                 style={{ display: 'inline-block', width: '30%' }}
                 errorText={errors.value} />
-              <span className={styles['percent-text']}>
-                %
-            </span>
             </div>
           </div>
           <div className={styles['discount-option-control']}>
@@ -73,6 +98,7 @@ const DiscountDetailForm = ({ discount, items, errors, onChange }) => {
               </div>
             </div>
           </div>
+
           <div className={styles['discount-option-control']}>
             <div className="row">
               {displayItemSelection &&
@@ -103,7 +129,7 @@ const DiscountDetailForm = ({ discount, items, errors, onChange }) => {
         </div>
         <div className="col-md-offset-1 col-sm-4">
           <div className={styles['detail-icon']}>
-            <i className="material-icons" style={{ fontSize: '124px' }}>account_balance</i>
+            <i className="material-icons" style={{ fontSize: '124px' }}>play_for_work</i>
           </div>
         </div>
       </div>